Write "Main" form logic and Redesign forms- check commit description.
* Connect to database.
* Write "Main" form logic.
* Retrieve data from Database
* Write "Delete" button logic.
* Redesign "Add" and "Edit" form.

+++ Database/DDL.sql
@@ -1,5 +1,5 @@
 CREATE TABLE question_type (
-    type_id tinyint PRIMARY KEY,
+    type_id int PRIMARY KEY,
     type_name nvarchar(255) NOT NULL
 );
 
@@ -7,12 +7,12 @@ CREATE TABLE question (
     question_id int IDENTITY (1, 1) PRIMARY KEY,
     question_text nvarchar(255) NOT NULL,
     question_order int NOT NULL CHECK (question_order >= 1),
-    type_id tinyint NOT NULL FOREIGN KEY REFERENCES question_type (type_id) ON DELETE CASCADE ON UPDATE CASCADE
+    type_id int NOT NULL FOREIGN KEY REFERENCES question_type (type_id) ON DELETE CASCADE ON UPDATE CASCADE
 );
 
 CREATE TABLE star_question (
     question_id int PRIMARY KEY FOREIGN KEY REFERENCES question (question_id) ON DELETE CASCADE ON UPDATE CASCADE,
-    num_of_stars tinyint DEFAULT 1 CHECK (
+    num_of_stars int DEFAULT 1 CHECK (
         num_of_stars >= 1
         AND num_of_stars <= 10
     )
@@ -20,7 +20,7 @@ CREATE TABLE star_question (
 
 CREATE TABLE smiley_question (
     question_id int PRIMARY KEY FOREIGN KEY REFERENCES question (question_id) ON DELETE CASCADE ON UPDATE CASCADE,
-    num_of_faces tinyint DEFAULT 2 CHECK (
+    num_of_faces int DEFAULT 2 CHECK (
         num_of_faces >= 2
         AND num_of_faces <= 5
     )
@@ -28,8 +28,8 @@ CREATE TABLE smiley_question (
 
 CREATE TABLE slider_question (
     question_id int PRIMARY KEY FOREIGN KEY REFERENCES question (question_id) ON DELETE CASCADE ON UPDATE CASCADE,
-    start_value tinyint DEFAULT 0,
-    end_value tinyint DEFAULT 100,
+    start_value int DEFAULT 0,
+    end_value int DEFAULT 100,
     start_value_caption nvarchar(255) NOT NULL,
     end_value_caption nvarchar(255) NOT NULL,
     CONSTRAINT check_value CHECK (
